requirements.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
Django==2.2.10
django-admin-sortable==2.2
django-celery-results==1.1.2
django-ckeditor==5.8.0
django-ckeditor-filebrowser-filer==0.3.0
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-filer==1.6.0
django-filter==2.2.0
django-js-asset==1.2.2
django-model-utils==3.2.0
django-mptt==0.10.0
django-polymorphic==2.0.3
django-redis==4.10.0
django-redis-cache==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.10.3
djangorestframework-camel-case==1.1.1
easy-thumbnails==2.7
gunicorn==20.0.4
importlib-metadata==0.23
jsmin==2.2.2
kombu==4.6.5
more-itertools==7.2.0
Pillow==6.1.0
psycopg2-binary==2.8.4
pytz==2019.2
redis==3.3.11
six==1.12.0
sorl-thumbnail==12.5.0
SQLAlchemy==1.3.10
twitter==1.18.0
Unidecode==1.0.23
vine==1.3.0
zipp==0.6.0
python-dotenv==0.12.0
